Backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
passlib[bcrypt]==1.7.4
python-jose==3.3.0
pydantic==2.5.0
python-multipart==0.0.6
ultralytics==8.0.228
python-dotenv==1.0.0
imageio==2.33.1
imageio-ffmpeg==0.4.9
opencv-python==4.8.1.78
numpy==1.24.3
torch==2.1.0
torchvision==0.16.0
keras==2.14.0
tensorflow==2.14.0
requests==2.31.0
Pillow==10.1.0